replaced long int usage with more reliably 64 bit long long

+++ GameMath/include/GameMath.h
@@ -362,7 +362,7 @@ class Xorshift64
 {
     public:
         Xorshift64();
-        Xorshift64(unsigned long int seed);
+        Xorshift64(unsigned long long seed);
         virtual ~Xorshift64();
         unsigned int   nextInt();
         unsigned int   nextInt(int mod);
@@ -373,25 +373,25 @@ class Xorshift64
         float nextFloat();
         double nextDouble();
         bool nextBool();
-        unsigned long int getSeed() const;
-        void setSeed(unsigned long int seed);
-        unsigned int getFromSeed(unsigned long int seed) const;
+        unsigned long long getSeed() const;
+        void setSeed(unsigned long long seed);
+        unsigned int getFromSeed(unsigned long long seed) const;
     protected:
     private:
-        const static unsigned long int MAXLONG = 0xffffffffffffffff;
-        unsigned long int seed;
-        unsigned long int val;
+        const static unsigned long long MAXLONG = 0xffffffffffffffff;
+        unsigned long long seed;
+        unsigned long long val;
 };
 
 
 class SpatialRandom {
     public:
         SpatialRandom();
-        SpatialRandom(unsigned long int seed);
+        SpatialRandom(unsigned long long seed);
         virtual ~SpatialRandom();
-        unsigned long int longFor(int x, int y, int z, int t);
-        unsigned long int longFor(int x, int y, int z);
-        unsigned long int longFor(int x, int y);
+        unsigned long long longFor(int x, int y, int z, int t);
+        unsigned long long longFor(int x, int y, int z);
+        unsigned long long longFor(int x, int y);
         float floatFor(int x, int y, int z, int t);
         float floatFor(int x, int y, int z);
         float floatFor(int x, int y);
@@ -401,13 +401,13 @@ class SpatialRandom {
         Xorshift64 xorshift64For(int x, int y, int z, int t);
         Xorshift64 xorshift64For(int x, int y, int z);
         Xorshift64 xorshift64For(int x, int y);
-        unsigned long int getSeed() const;
-        void setSeed(unsigned long int seed);
+        unsigned long long getSeed() const;
+        void setSeed(unsigned long long seed);
     protected:
     private:
-        const static unsigned long int MAXLONG = 0xffffffffffffffff;
-        unsigned long int seed;
-        unsigned long int val;
+        const static unsigned long long MAXLONG = 0xffffffffffffffff;
+        unsigned long long seed;
+        unsigned long long val;
 };
